Event listener - OpenTelemetry integration
