--- presto-db2fori/src/main/java/com/facebook/presto/plugin/db2fori/Db2foriPlugin.java
@@ -1,25 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.facebook.presto.plugin.db2fori;
-
-import com.facebook.presto.plugin.jdbc.JdbcPlugin;
-
-public class Db2foriPlugin
-        extends JdbcPlugin
-{
-    public Db2foriPlugin()
-    {
-        super("db2fori", new Db2foriClientModule());
-    }
-}
