+++ presto-event-listener/src/main/java/com/facebook/presto/eventlistener/QueryEventListenerPlugin.java
@@ -1,0 +1,29 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.eventlistener;
+
+import com.facebook.presto.spi.Plugin;
+import com.facebook.presto.spi.eventlistener.EventListenerFactory;
+
+import java.util.Collections;
+
+public final class QueryEventListenerPlugin
+        implements Plugin
+{
+    public Iterable getEventListenerFactories()
+    {
+        EventListenerFactory listenerFactory = new QueryEventListenerFactory();
+        return Collections.singletonList(listenerFactory);
+    }
+}
