+++ presto-event-listener/src/resources/META-INF/services/com.facebook.presto.spi.Plugin
@@ -1,0 +1,1 @@
+com.facebook.presto.eventlistener.QueryEventListenerPlugin
