+++ presto-event-listener/src/test/resources/onlyQueryCreatedEvents.properties
@@ -22,7 +22,7 @@ appender.rolling.policies.size.size=100MB
 appender.rolling.strategy.type = DefaultRolloverStrategy
 appender.rolling.strategy.max = 5
 
-logger.rolling.name = io.ahana.eventplugin
+logger.rolling.name = com.facebook.presto.eventlistener
 logger.rolling.level = debug
 logger.rolling.additivity = false
 logger.rolling.appenderRef.rolling.ref = RollingFile
