+++ presto-main/etc/config.properties
@@ -26,6 +26,10 @@ scheduler.http-client.idle-timeout=1m
 query.client.timeout=5m
 query.min-expire-age=30m
 
+tracing.tracer-type=otel
+tracing.enable-distributed-tracing=true
+tracing.distributed-tracing-mode=always_trace
+
 plugin.bundles=\
   ../presto-blackhole/pom.xml,\
   ../presto-memory/pom.xml,\
@@ -50,8 +54,11 @@ plugin.bundles=\
   ../presto-cluster-ttl-providers/pom.xml,\
   ../presto-node-ttl-fetchers/pom.xml,\
   ../presto-hive-function-namespace/pom.xml,\
+  ../presto-event-listener/pom.xml,\
+  ../presto-open-telemetry/pom.xml,\
   ../presto-delta/pom.xml,\
   ../presto-hudi/pom.xml
 
 presto.version=testversion
 node-scheduler.include-coordinator=true
+coordinator=true
