--- presto-teradata/pom.xml
@@ -1,127 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>presto-root</artifactId>
-    <groupId>com.facebook.presto</groupId>
-    <version>0.286</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-
-  <artifactId>presto-teradata</artifactId>
-  <description>Presto - teradata Connector</description>
-  <packaging>presto-plugin</packaging>
-
-  <properties>
-    <!--suppress UnresolvedMavenProperty -->
-    <air.main.basedir>${project.parent.basedir}</air.main.basedir>
-    <air.check.skip-enforcer>true</air.check.skip-enforcer>
-    <air.check.skip-dependency>true</air.check.skip-dependency>
-  </properties>
-
-  <dependencies>
-    <!-- Presto SPI -->
-    <dependency>
-      <groupId>com.facebook.presto</groupId>
-      <artifactId>presto-spi</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>com.facebook.presto</groupId>
-      <artifactId>presto-base-jdbc</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.facebook.presto</groupId>
-      <artifactId>presto-common</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>io.airlift</groupId>
-      <artifactId>slice</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>io.airlift</groupId>
-      <artifactId>units</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>com.fasterxml.jackson.core</groupId>
-      <artifactId>jackson-annotations</artifactId>
-      <scope>provided</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>org.testng</groupId>
-      <artifactId>testng</artifactId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.google.inject</groupId>
-      <artifactId>guice</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.facebook.airlift</groupId>
-      <artifactId>configuration</artifactId>
-    </dependency>
-
-    <!-- for testing -->
-    <dependency>
-      <groupId>com.facebook.presto</groupId>
-      <artifactId>presto-testng-services</artifactId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>com.facebook.presto</groupId>
-      <artifactId>presto-main</artifactId>
-      <scope>test</scope>
-    </dependency>
-
-    <dependency>
-      <groupId>com.facebook.airlift</groupId>
-      <artifactId>security</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>com.teradata.jdbc</groupId>
-      <artifactId>terajdbc</artifactId>
-      <version>20.00.00.11</version>
-    </dependency>
-
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.13.1</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.basepom.maven</groupId>
-        <artifactId>duplicate-finder-maven-plugin</artifactId>
-        <configuration>
-          <ignoredClassPatterns>
-            <ignoredClassPattern>META-INF.versions.9.module-info</ignoredClassPattern>
-            <ignoredClassPattern> module-info</ignoredClassPattern>
-          </ignoredClassPatterns>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
-</project>
